requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
python-docx==0.8.11
groq==0.4.1
numpy==1.24.3
nltk==3.8.1
scikit-learn==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
requests==2.31.0